Add Problems Solved
