+++ practice/c++/Input and Output/file.cpp
@@ -1,0 +1,19 @@
+#include <cmath>
+#include <cstdio>
+#include <vector>
+#include <iostream>
+#include <algorithm>
+using namespace std;
+
+
+int main() {
+    int a, b, c, total;
+
+    cin >> a >> b >> c;
+
+    total = a + b + c;
+
+    cout << total;
+
+    return 0;
+}
