+++ practice/python/Hello World/file.py
@@ -1,0 +1,2 @@
+if __name__ == '__main__':
+    print("Hello, World!")
