+++ practice/sql/Japanese Cities Attributes/file.sql
@@ -1,0 +1,1 @@
+SELECT * FROM CITY WHERE COUNTRYCODE = 'JPN';
