+++ practice/sql/Japanese Cities Names/file.sql
@@ -1,0 +1,1 @@
+SELECT NAME FROM CITY WHERE COUNTRYCODE = 'JPN';
