+++ practice/sql/Revising the Select Query I/file.sql
@@ -1,0 +1,2 @@
+SELECT * FROM CITY
+WHERE COUNTRYCODE = 'USA' AND POPULATION > 100000;
