+++ practice/sql/Revising the Select Query II/file.sql
@@ -1,0 +1,2 @@
+SELECT NAME FROM CITY
+WHERE COUNTRYCODE = 'USA' AND POPULATION > 120000;
