+++ practice/sql/Select All/file.sql
@@ -1,0 +1,1 @@
+SELECT * FROM CITY;
