+++ practice/sql/Select By ID/file.sql
@@ -1,0 +1,1 @@
+SELECT * FROM CITY WHERE ID = 1661;
