+++ practice/sql/Weather Observation Station 1/file.sql
@@ -1,0 +1,1 @@
+SELECT CITY, STATE FROM STATION;
