+++ practice/sql/Weather Observation Station 3/file.sql
@@ -1,0 +1,1 @@
+SELECT DISTINCT CITY FROM STATION WHERE ID % 2 = 0;
