--- run/c++/CMakeLists.txt
@@ -1,6 +1,0 @@
-# cmake_minimum_required(VERSION <specify CMake version here>)
-project(hackerrank-solutions)
-
-set(CMAKE_CXX_STANDARD 14)
-
-add_executable(hackerrank-solutions main.cpp)
