--- run/c++/in
@@ -1,1 +1,0 @@
-1 2 3 4 5
