--- run/c++/out
@@ -1,1 +1,0 @@
-10 14
