+++ run/cpp/CMakeLists.txt
@@ -1,0 +1,6 @@
+cmake_minimum_required(VERSION 3.24)
+project(cpp)
+
+set(CMAKE_CXX_STANDARD 14)
+
+add_executable(cpp main.cpp)
