+++ run/cpp/in
@@ -1,0 +1,1 @@
+1 2 7
