+++ run/cpp/main.cpp
@@ -1,0 +1,29 @@
+//C++
+#include <bits/stdc++.h>
+
+//C
+//#include <stdio.h>
+//#include <string.h> (strlen)
+//#include <cstdio>
+
+//Quitar para C
+using namespace std;
+
+int main() {
+
+    freopen("..\\in", "r", stdin);
+    freopen("..\\out", "w", stdout);
+
+    ios_base::sync_with_stdio(false);
+    cin.tie(nullptr);
+
+    int a, b, c, total;
+
+    cin >> a >> b >> c;
+
+    total = a + b + c;
+
+    cout << total;
+
+    return 0;
+}
