+++ run/cpp/out
@@ -1,0 +1,1 @@
+10
